chore : removeListeners 메서드 Int 타입 파라미터로 변경

+++ android/app/src/main/java/com/fitpet/StepSensorModule.kt
@@ -46,5 +46,5 @@ class StepSensorModule(private val reactContext: ReactApplicationContext) :
     fun addListener(eventName: String) {}
 
     @ReactMethod
-    fun removeListeners(count: Double) {}
+    fun removeListeners(count: Int) {}
 }
